Fix tests

+++ src/metatensor/models/experimental/alchemical_model/tests/test_torchscript.py
@@ -20,7 +20,7 @@ def test_torchscript():
             )
         },
     )
-    alchemical_model = Model(capabilities, DEFAULT_HYPERS["model"]).to(torch.float64)
+    alchemical_model = Model(capabilities, DEFAULT_HYPERS["model"])
     torch.jit.script(
         alchemical_model, {"energy": alchemical_model.capabilities.outputs["energy"]}
     )
@@ -39,7 +39,7 @@ def test_torchscript_save():
             )
         },
     )
-    alchemical_model = Model(capabilities, DEFAULT_HYPERS["model"]).to(torch.float64)
+    alchemical_model = Model(capabilities, DEFAULT_HYPERS["model"])
     torch.jit.save(
         torch.jit.script(
             alchemical_model,
