+++ src/metatensor/models/experimental/pet/tests/test_functionality.py
@@ -1,5 +1,4 @@
 import ase
-import torch
 from metatensor.torch.atomistic import (
     MetatensorAtomisticModel,
     ModelCapabilities,
@@ -29,9 +28,7 @@ def test_prediction_subset():
         supported_devices=["cpu"],
     )
 
-    model = Model(capabilities, DEFAULT_HYPERS["ARCHITECTURAL_HYPERS"]).to(
-        torch.float64
-    )
+    model = Model(capabilities, DEFAULT_HYPERS["ARCHITECTURAL_HYPERS"])
     structure = ase.Atoms("O2", positions=[[0.0, 0.0, 0.0], [0.0, 0.0, 1.0]])
     system = systems_to_torch(structure)
     system = get_system_with_neighbors_lists(system, model.requested_neighbors_lists())
